Moved test data to its own directory
